refactor: extract position delete helpers and simplify pk-index writer
