+++ src/stream/Cargo.toml
@@ -44,7 +44,6 @@ memcomparable = "0.2"
 moka = { version = "0.12.0", features = ["sync"] }
 multimap = "0.10"
 parking_lot = { workspace = true }
-parquet = { workspace = true }
 paste = "1"
 pin-project = "1"
 prometheus = { version = "0.14", features = ["process"] }
