+++ src/stream/src/executor/iceberg_with_pk_index/position_delete_handler_impl.rs
@@ -12,44 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::collections::HashMap as StdHashMap;
-use std::sync::Arc;
 use std::time::Duration;
 
 use anyhow::{Context, anyhow};
-use futures::StreamExt;
 use hashbrown::HashMap;
 use hashbrown::hash_map::Entry;
-use iceberg::arrow::schema_to_arrow_schema;
 use iceberg::delete_vector::DeleteVector;
-use iceberg::io::FileIO;
-use iceberg::puffin::{CompressionCodec, PuffinReader, PuffinWriter};
 use iceberg::spec::{
-    DataContentType, DataFile, DataFileBuilder, DataFileFormat, FormatVersion, ManifestContentType,
+    DataContentType, DataFile, DataFileFormat, FormatVersion, ManifestContentType,
     SerializedDataFile,
 };
 use iceberg::table::Table;
-use iceberg::writer::base_writer::position_delete_file_writer::POSITION_DELETE_SCHEMA;
 use iceberg::writer::file_writer::location_generator::{
-    DefaultFileNameGenerator, DefaultLocationGenerator, FileNameGenerator, LocationGenerator,
+    DefaultFileNameGenerator, DefaultLocationGenerator,
 };
-use iceberg::writer::file_writer::{
-    FileWriter, FileWriterBuilder, ParquetWriter, ParquetWriterBuilder,
-};
-use parquet::arrow::{ParquetRecordBatchStreamBuilder, ProjectionMask};
-use parquet::file::properties::WriterProperties;
-use risingwave_common::array::arrow::arrow_array_iceberg::{
-    Array, ArrayRef, Int64Array, RecordBatch, StringArray,
-};
-use risingwave_common::array::arrow::arrow_schema_iceberg::SchemaRef as ArrowSchemaRef;
 use risingwave_common::bitmap::Bitmap;
 use risingwave_common::id::SinkId;
 use risingwave_connector::sink::iceberg::{
-    IcebergConfig, IcebergPositionDeleteMergerCommitResult, PARQUET_CREATED_BY,
-    resolve_partition_type, serialize_data_files_default_spec,
+    IcebergConfig, IcebergPositionDeleteCommitResult, read_position_deletes_from_file,
+    resolve_partition_type, serialize_data_files_default_spec, write_dv_puffin_file,
+    write_parquet_position_delete_file,
 };
 use risingwave_connector::sink::{Result as SinkResult, SinkError};
-use risingwave_connector::source::iceberg::parquet_file_handler::ParquetFileReader;
 use risingwave_pb::connector_service::SinkMetadata;
 use risingwave_pb::id::ActorId;
 use risingwave_rpc_client::MetaClient;
@@ -59,11 +43,6 @@ use uuid::Uuid;
 
 use super::position_delete_merger::PositionDeleteHandler;
 use super::position_delete_staging::StagingVersion;
-
-/// Puffin blob property for deletion vector cardinality.
-const DELETION_VECTOR_PROPERTY_CARDINALITY: &str = "cardinality";
-/// Puffin blob property for referenced data file path.
-const DELETION_VECTOR_PROPERTY_REFERENCED_DATA_FILE: &str = "referenced-data-file";
 
 /// Real implementation of [`PositionDeleteHandler`] using the iceberg-rust crate.
 ///
@@ -92,7 +71,7 @@ struct SeededState {
     table: Table,
     location_generator: DefaultLocationGenerator,
     /// File-name generator for V3 Puffin deletion vector files.
-    file_name_generator: DefaultFileNameGenerator,
+    puffin_file_name_generator: DefaultFileNameGenerator,
     /// File-name generator for V2 Parquet position-delete files.
     parquet_file_name_generator: DefaultFileNameGenerator,
     schema_id: i32,
@@ -171,11 +150,6 @@ impl PositionDeleteHandlerImpl {
         let pending = std::mem::take(&mut self.pending);
         let state = self.await_ready().await?;
 
-        // V3 tables use Puffin deletion vectors; V2 tables use file-scoped Parquet
-        // position-delete files. The internal representation is a roaring bitmap
-        // (`DeleteVector`) in both cases; only the on-disk format differs.
-        let use_puffin = state.format_version >= FormatVersion::V3;
-
         let mut delete_files = Vec::with_capacity(pending.len());
         let mut overwrite_files = Vec::new();
 
@@ -188,25 +162,9 @@ impl PositionDeleteHandlerImpl {
                     let existing_file = entry
                         .current_file()
                         .expect("needs_load implies current_file is Some");
-                    let dv = match existing_file.file_format() {
-                        DataFileFormat::Puffin => {
-                            read_dv_positions_from_data_file(state.table.file_io(), &existing_file)
-                                .await?
-                        }
-                        DataFileFormat::Parquet => {
-                            read_v2_positions_from_delete_file(
-                                state.table.file_io(),
-                                &existing_file,
-                            )
-                            .await?
-                        }
-                        other => {
-                            return Err(SinkError::Iceberg(anyhow!(
-                                "unexpected delete file format {:?}",
-                                other
-                            )));
-                        }
-                    };
+                    let dv = read_position_deletes_from_file(state.table.file_io(), existing_file)
+                        .await
+                        .map_err(SinkError::Iceberg)?;
                     entry.set_loaded_delete_vector(dv);
                 }
                 entry.plan_write(new_positions)
@@ -215,25 +173,33 @@ impl PositionDeleteHandlerImpl {
                 overwrite_files.push(overwrite);
             }
 
+            // V3 tables use Puffin deletion vectors; V2 tables use file-scoped Parquet
+            // position-delete files. The internal representation is a roaring bitmap
+            // (`DeleteVector`) in both cases; only the on-disk format differs.
+            let use_puffin = state.format_version >= FormatVersion::V3;
             let new_file = if use_puffin {
                 write_dv_puffin_file(
-                    &state.file_name_generator,
-                    &state.location_generator,
                     &state.table,
+                    &state.location_generator,
+                    &state.puffin_file_name_generator,
                     data_file_path.clone(),
                     &plan.merged,
+                    None,
                 )
-                .await?
+                .await
+                .map_err(SinkError::Iceberg)?
             } else {
                 write_parquet_position_delete_file(
-                    &state.parquet_file_name_generator,
-                    &state.location_generator,
                     &state.table,
+                    &state.location_generator,
+                    &state.parquet_file_name_generator,
                     &config,
                     data_file_path.clone(),
                     &plan.merged,
+                    None,
                 )
-                .await?
+                .await
+                .map_err(SinkError::Iceberg)?
             };
 
             state
@@ -250,7 +216,7 @@ impl PositionDeleteHandlerImpl {
         let delete_files = serialize_delete_files(&state.table, delete_files)?;
         let overwrite_files = serialize_overwrite_files(&state.table, overwrite_files)?;
 
-        let sink_metadata = SinkMetadata::try_from(&IcebergPositionDeleteMergerCommitResult {
+        let sink_metadata = SinkMetadata::try_from(&IcebergPositionDeleteCommitResult {
             schema_id: state.schema_id,
             partition_spec_id: state.partition_spec_id,
             delete_files,
@@ -258,153 +224,6 @@ impl PositionDeleteHandlerImpl {
         })?;
         Ok(Some(sink_metadata))
     }
-}
-
-/// Writes `delete_vector` as a single V3 Puffin deletion vector blob referencing
-/// `data_file_path`, and returns its [`DataFile`] metadata with
-/// `referenced_data_file` set. The delete file is written with an empty partition;
-/// the coordinator backfills the real partition in a later stage.
-async fn write_dv_puffin_file(
-    file_name_generator: &DefaultFileNameGenerator,
-    location_generator: &DefaultLocationGenerator,
-    table: &Table,
-    data_file_path: String,
-    delete_vector: &DeleteVector,
-) -> SinkResult<DataFile> {
-    let file_name = file_name_generator.generate_file_name();
-    // Empty partition: the coordinator backfills the real partition.
-    let location = location_generator.generate_location(None, &file_name);
-    let output_file = table.file_io().new_output(&location)?;
-    let mut writer = PuffinWriter::new(&output_file, StdHashMap::new(), false).await?;
-
-    let cardinality = delete_vector.len();
-    let properties = StdHashMap::from([
-        (
-            DELETION_VECTOR_PROPERTY_CARDINALITY.to_owned(),
-            cardinality.to_string(),
-        ),
-        (
-            DELETION_VECTOR_PROPERTY_REFERENCED_DATA_FILE.to_owned(),
-            data_file_path.clone(),
-        ),
-    ]);
-    let blob = delete_vector.to_puffin_blob(properties)?;
-    writer.add(blob, CompressionCodec::None).await?;
-
-    let result = writer.close_with_metadata().await?;
-    let blob_metadata = result
-        .blobs_metadata
-        .first()
-        .ok_or_else(|| SinkError::Iceberg(anyhow!("blob metadata should be present")))?;
-
-    let mut builder = DataFileBuilder::default();
-    builder
-        .content(DataContentType::PositionDeletes)
-        .file_path(location)
-        .file_format(DataFileFormat::Puffin)
-        .record_count(cardinality)
-        .file_size_in_bytes(result.file_size_in_bytes)
-        .referenced_data_file(Some(data_file_path))
-        .content_offset(Some(blob_metadata.offset() as i64))
-        .content_size_in_bytes(Some(blob_metadata.length() as i64));
-    builder.build().map_err(|err| {
-        SinkError::Iceberg(anyhow!(err).context("Failed to build deletion vector file metadata"))
-    })
-}
-
-/// Writes `delete_vector` as a single file-scoped V2 Parquet position-delete file,
-/// and returns the resulting [`DataFile`] metadata. The produced file references
-/// exactly one data file (`data_file_path`) via `referenced_data_file` and is
-/// written with an empty partition; the coordinator backfills the real partition
-/// in a later stage.
-///
-/// File-scoped deletes reference exactly one data file, so this writes the file
-/// with a plain [`ParquetWriter`] rather than the sink's rolling position-delete
-/// writer stack.
-async fn write_parquet_position_delete_file(
-    file_name_generator: &DefaultFileNameGenerator,
-    location_generator: &DefaultLocationGenerator,
-    table: &Table,
-    config: &IcebergConfig,
-    data_file_path: String,
-    delete_vector: &DeleteVector,
-) -> SinkResult<DataFile> {
-    let file_name = file_name_generator.generate_file_name();
-    // Empty partition: the coordinator backfills the real partition.
-    let location = location_generator.generate_location(None, &file_name);
-    let output_file = table.file_io().new_output(&location)?;
-
-    let parquet_writer_properties = WriterProperties::builder()
-        .set_compression(config.get_parquet_compression())
-        .set_max_row_group_bytes(config.write_parquet_max_row_group_bytes())
-        .set_created_by(PARQUET_CREATED_BY.to_owned())
-        .build();
-    let mut writer = ParquetWriterBuilder::new(
-        parquet_writer_properties,
-        POSITION_DELETE_SCHEMA.clone().into(),
-    )
-    .build(output_file)
-    .await?;
-
-    // The position-delete schema is `(file_path, pos)` with reserved field IDs;
-    // derive the matching Arrow schema so the written column field IDs line up.
-    let arrow_schema: ArrowSchemaRef = Arc::new(schema_to_arrow_schema(&POSITION_DELETE_SCHEMA)?);
-
-    const POSITION_DELETE_WRITE_CHUNK_SIZE: usize = 1024;
-    let mut positions: Vec<i64> = Vec::with_capacity(POSITION_DELETE_WRITE_CHUNK_SIZE);
-    for pos in delete_vector.iter() {
-        positions.push(pos as i64);
-        if positions.len() == POSITION_DELETE_WRITE_CHUNK_SIZE {
-            write_position_delete_chunk(
-                &mut writer,
-                &arrow_schema,
-                &data_file_path,
-                std::mem::take(&mut positions),
-            )
-            .await?;
-            positions.reserve(POSITION_DELETE_WRITE_CHUNK_SIZE);
-        }
-    }
-    if !positions.is_empty() {
-        write_position_delete_chunk(&mut writer, &arrow_schema, &data_file_path, positions).await?;
-    }
-
-    let data_files = writer.close().await?;
-    // `close` will yield exactly one builder here.
-    let [mut builder] = data_files.try_into().map_err(|_| {
-        SinkError::Iceberg(anyhow!(
-            "position-delete writer produced invalid file count for {data_file_path}"
-        ))
-    })?;
-
-    // `ParquetWriter` builds the file as `DataContentType::Data` with an empty
-    // partition; override those for a file-scoped V2 position-delete file and
-    // attach `referenced_data_file`.
-    builder
-        .content(DataContentType::PositionDeletes)
-        .referenced_data_file(Some(data_file_path));
-    builder.build().map_err(|err| {
-        SinkError::Iceberg(anyhow!(err).context("Failed to build position-delete file metadata"))
-    })
-}
-
-/// Writes one chunk of `positions` as a `(file_path, pos)` batch into `writer`.
-/// Every row shares `data_file_path` because the delete file is file-scoped.
-async fn write_position_delete_chunk(
-    writer: &mut ParquetWriter,
-    arrow_schema: &ArrowSchemaRef,
-    data_file_path: &str,
-    positions: Vec<i64>,
-) -> SinkResult<()> {
-    let path_column: ArrayRef = Arc::new(StringArray::from_iter_values(std::iter::repeat_n(
-        data_file_path,
-        positions.len(),
-    )));
-    let pos_column: ArrayRef = Arc::new(Int64Array::from(positions));
-    let batch = RecordBatch::try_new(arrow_schema.clone(), vec![path_column, pos_column])
-        .map_err(|e| SinkError::Iceberg(anyhow!(e)))?;
-    writer.write(&batch).await?;
-    Ok(())
 }
 
 #[async_trait::async_trait]
@@ -434,7 +253,7 @@ impl PositionDeleteHandler for PositionDeleteHandlerImpl {
                 // 3. Derive per-table state + seed staging (shard-filtered).
                 let location_generator = DefaultLocationGenerator::new(table.metadata().clone())?;
                 let uuid_suffix = Uuid::now_v7();
-                let file_name_generator = DefaultFileNameGenerator::new(
+                let puffin_file_name_generator = DefaultFileNameGenerator::new(
                     actor_id.to_string(),
                     Some(format!("delvec-{}", uuid_suffix)),
                     DataFileFormat::Puffin,
@@ -454,7 +273,7 @@ impl PositionDeleteHandler for PositionDeleteHandlerImpl {
                 Ok(SeededState {
                     table,
                     location_generator,
-                    file_name_generator,
+                    puffin_file_name_generator,
                     parquet_file_name_generator,
                     schema_id,
                     partition_spec_id,
@@ -590,91 +409,4 @@ fn serialize_overwrite_files(
             Ok(SerializedDataFile::try_from(f, pt, format_version)?)
         })
         .collect()
-}
-
-async fn read_dv_positions_from_data_file(
-    file_io: &FileIO,
-    data_file: &DataFile,
-) -> SinkResult<DeleteVector> {
-    let blob_offset = data_file.content_offset().ok_or_else(|| {
-        SinkError::Iceberg(anyhow!(
-            "DV file {} missing content_offset for referenced data file {:?}",
-            data_file.file_path(),
-            data_file.referenced_data_file()
-        ))
-    })?;
-    let blob_length = data_file.content_size_in_bytes().ok_or_else(|| {
-        SinkError::Iceberg(anyhow!(
-            "DV file {} missing content_size_in_bytes for referenced data file {:?}",
-            data_file.file_path(),
-            data_file.referenced_data_file()
-        ))
-    })?;
-
-    let input_file = file_io.new_input(data_file.file_path())?;
-    let puffin_reader = PuffinReader::new(input_file);
-    let file_metadata = puffin_reader.file_metadata().await?;
-    let blob_metadata = file_metadata
-        .blobs()
-        .iter()
-        .find(|blob| blob.offset() == blob_offset as u64 && blob.length() == blob_length as u64)
-        .ok_or_else(|| {
-            SinkError::Iceberg(anyhow!(
-                "DV blob metadata not found in {} at offset={} length={}",
-                data_file.file_path(),
-                blob_offset,
-                blob_length
-            ))
-        })?;
-    let blob = puffin_reader.blob(blob_metadata).await?;
-
-    let delete_vector = DeleteVector::from_puffin_blob(blob)?;
-    Ok(delete_vector)
-}
-
-/// Reads the positions stored in a V2 Parquet position-delete file into a
-/// [`DeleteVector`]. The file's schema is `(file_path, pos)`. Callers only invoke
-/// this after the entry's `referenced_data_file` already matched the target data
-/// file, and the files we write are file-scoped (every row shares one
-/// `file_path`), so the `file_path` column is redundant here: we project only the
-/// `pos` column and read every value.
-async fn read_v2_positions_from_delete_file(
-    file_io: &FileIO,
-    data_file: &DataFile,
-) -> SinkResult<DeleteVector> {
-    let input_file = file_io.new_input(data_file.file_path())?;
-    let metadata = input_file.metadata().await?;
-    let reader = input_file.reader().await?;
-    let parquet_reader = ParquetFileReader::new(metadata, reader);
-    let builder = ParquetRecordBatchStreamBuilder::new(parquet_reader).await?;
-    // The position-delete schema fixes column order to (file_path, pos); project
-    // only the `pos` leaf (column index 1) so the file_path column is never decoded.
-    let projection = ProjectionMask::leaves(builder.parquet_schema(), [1]);
-    let mut stream = builder.with_projection(projection).build()?;
-
-    let mut delete_vector = DeleteVector::default();
-    while let Some(batch) = stream.next().await {
-        let batch = batch?;
-        // Only the projected `pos` column is present in the batch.
-        let positions = batch.columns()[0]
-            .as_any()
-            .downcast_ref::<Int64Array>()
-            .ok_or_else(|| {
-                SinkError::Iceberg(anyhow!(
-                    "position-delete pos column should be an Int64Array"
-                ))
-            })?;
-
-        for pos in positions {
-            let Some(pos) = pos else {
-                return Err(SinkError::Iceberg(anyhow!(
-                    "null value in position-delete file {}",
-                    data_file.file_path()
-                )));
-            };
-            delete_vector.insert(pos as u64);
-        }
-    }
-
-    Ok(delete_vector)
 }
